[refactor]ボタンのロジックとUIを分離
