+++ src/components/button.jsx
@@ -1,0 +1,9 @@
+const Button = ({ handleTask, children }) => {
+  return (
+    <button onClick={() => handleTask()}>
+     {children}
+    </button>
+  );
+};
+
+export default Button;
